brewmmer/requirements.txt
Django==1.9.4
django-bootstrap3==7.0.1
django-nested-admin==2.2.6
django-nested-inline==0.3.6
django-smart-selects==1.2.2
six==1.10.0
wheel==0.26.0